Enum plist_utils::errors::ParseError

pub enum ParseError {

    FileTooShort(usize),
    InvalidAsciiString(Vec<u8>),
    InvalidDataOffset(usize),
    InvalidDate(Vec<u8>),
    InvalidFloat(u8, Vec<u8>),
    InvalidHeaderTag,
    InvalidInteger(u8, Vec<u8>),
    InvalidNullOrBool(u8),
    InvalidPrefix(u8),
    InvalidUtf16String(Vec<u8>),
    MissingDateMarker(u8),
    NomError,
    UnsupportedVersion,

}

Variants

FileTooShort(usize)

Tuple Fields

0: usize

Input ended too soon.

InvalidAsciiString(Vec<u8>)

Tuple Fields

0: Vec<u8>

The byte sequence cannot represent a valid ASCII string.

InvalidDataOffset(usize)

Tuple Fields

0: usize

An out of bounds error upon using one of the offsets derived from the buffer.

InvalidDate(Vec<u8>)

Tuple Fields

0: Vec<u8>

The serialized date object cannot be parsed.

Potential reasons include:

Buffer length doesn’t match 4 bytes.

Buffer doesn’t contain a valid f64

InvalidFloat(u8, Vec<u8>)

Tuple Fields

0: u8

1: Vec<u8>

The serialized floating point number could not be parsed.

Potential reasons include:

Wrong width (something other than 1, 2, or 4.

The number of bytes didn’t match what the width indicated.

The buffer could not be parsed into a floating point number.

InvalidHeaderTag

The file header was not recognized.

InvalidInteger(u8, Vec<u8>)

Tuple Fields

0: u8

1: Vec<u8>

The serialized integer could not be read. This is most likely due to the width not matching the number of bytes.

InvalidNullOrBool(u8)

Tuple Fields

0: u8

Expected a null or bool due to high 4 bits being 0x0, but could not match the low 4 bits with any of null (0x00), true (0x07), or false (0x08).

InvalidPrefix(u8)

Tuple Fields

0: u8

The prefix (4 bits) could not be recognized.

InvalidUtf16String(Vec<u8>)

Tuple Fields

0: Vec<u8>

The byte sequence cannot represent a valid UTF-16 string.

MissingDateMarker(u8)

Tuple Fields

0: u8

0x33 is the date marker. This is used when the parser sees a 0x30 in the high 4 bits, but sees something other than 0x03 in the lower 4 bits.

NomError

Generic nom error that has not yet been classified.

UnsupportedVersion

The property list version is not yet supported by the library.